Pull complex code out of a header
This may have made sense as all-inlined stuff at one time, but not any
more.  Also, the ValueReachesValue class is an internal implementation
detail that need not be visible to anyone else.
